.legacy/pyproject.toml
[build-system]
requires = ["setuptools>=61,<66", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "h5radiomics"
version = "0.1.1"
description = "Pipeline for radiomics feature extraction and analysis from H5-based WSI patches. Currently extracts intensity+texture based features only, and separately outputs cell segmentations."
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24,<2.0.0",
    "pandas>=1.5",
    "pyyaml",
    "h5py",
    "Pillow",
    "SimpleITK",
    "pyradiomics==3.0.1",
    "matplotlib",
    "tqdm",
    "shapely",
    "geopandas",
    "opencv-python",
    "gdown",
    "pyarrow",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
